progress, check TODO

+++ Proyecto/src/main/java/Main.java
@@ -3,6 +3,7 @@ import DBs.Skill;
 import DBs.Skillset;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 import java.util.ArrayList;
@@ -20,9 +21,10 @@ public class Main {
         LogManager.getLogManager().getLogger("").setLevel(Level.SEVERE);
 
         System.out.println("TESTEADORS = ");
-       removeSkill();
-        //crearEmpleado("test4","test4","test4");
 
+       //removeSkill();
+        //crearEmpleado("test4","test4","test4");
+    highestID();
     }
 
 
@@ -33,34 +35,55 @@ public class Main {
         transaction.commit();
         System.out.println("Usuario "+emp.getName()+ " creado con exito");
     }
+    public static Employee getEmployee(String nif){
+        Employee emp;
+        transaction.begin();
+        emp = (Employee) entityManager.createNativeQuery("SELECT * FROM EMPLOYEE WHERE NIF = ?" )
+                .setParameter(1, nif)
+                .getSingleResult();
+
+        transaction.commit();
+        return emp;
+    }
+
     public static Integer crearSkillsetTest(){
         int[] array = highestID();
         Skillset ss = new Skillset(array[0],array[1], 0,0);
-        transaction.begin();
-            entityManager.persist(ss) ; //TODO skill y nivel, conectar con skill
-        transaction.commit();
+        //transaction.begin();
+          //  entityManager.persist(ss) ; //TODO skill y nivel, conectar con skill
+        //transaction.commit();
         return array[1];
     }
 
 
     public static int[] highestID(){
-        ArrayList<Integer> test = new ArrayList<>() ;
-        ArrayList<Integer> test2 = new ArrayList<>() ;
 
-            Query query = entityManager.createNativeQuery(
-                    "SELECT * FROM SKILLSET", Skillset.class);
-            List<Skillset> res = query.getResultList();
-            for (Skillset i :  res) {
-                test.add(i.getSsId().intValue());
-                test2.add(i.getSkillset().intValue());
-            }
+        BigDecimal id1= (BigDecimal) entityManager.createNativeQuery("SELECT MAX(SS_ID) FROM SKILLSET")
+                .getSingleResult();
+        BigDecimal id2 = (BigDecimal) entityManager.createNativeQuery("SELECT MAX(SKILLSET) FROM SKILLSET")
+                .getSingleResult();
+
+
         int[] array = new int[2];
-        Integer max = iterator(test);
-        Integer max2 = iterator(test2);
-        array[0] =max+1;
-        array[1] = max2+1;
-        return (array);
+        array[0] =id1.intValue()+1;
+        array[1] = id2.intValue()+1;
 
+        return (array);
+        /**
+         *         ArrayList<Integer> test = new ArrayList<>() ;
+         *         ArrayList<Integer> test2 = new ArrayList<>() ;
+         Query query = entityManager.createNativeQuery(
+         "SELECT * FROM SKILLSET", Skillset.class);
+
+         List<Skillset> res = query.getResultList();
+         for (Skillset i :  res) {
+         test.add(i.getSsId().intValue());
+         test2.add(i.getSkillset().intValue());
+         }
+
+         Integer max = iterator(test);
+         Integer max2 = iterator(test2);
+         **/
     }
     public static void createSkill(String skill, String desc){
 
@@ -73,45 +96,70 @@ public class Main {
                 .setParameter(2, desc)
                 .executeUpdate();
 
-
-
         transaction.commit();
         System.out.println("AAAA CURBAAAAA");
     }
-    public static void removeSkill() {
-        int res = 5;
-        //transaction.begin();
+
+    public static Boolean skillExists(String skill){
+        transaction.begin();
+
+        Query query = entityManager.createNativeQuery("SELECT * FROM SKILL WHERE SKILL_NAME = ?", Skill.class)
+                .setParameter(1, skill);
+
+        try {
+            Skill results = (Skill) query.getSingleResult();
+            System.out.println("Existe esa skill");
+            transaction.commit();
+            return true;
+            // entityManager.remove(results);
+
+        }catch (NoResultException e){
+            transaction.commit();
+            System.out.println("No existe esa skill");
+            return false;
+
+        }
+
+    }
+
+
+    public static void removeSkill(String skill) {
 
         //pasar la skil a liminar (leer skill_name y obtener su id)
+if(skillExists(skill)) {
 
-            transaction.begin();
-            System.out.println("RESU = ");
-             String skill = sc.nextLine();
-            Query query = entityManager.createNativeQuery("SELECT * FROM SKILL WHERE SKILL_NAME = ?", Skill.class)
-                    .setParameter(1, skill);
-            try {
-                Skill results = (Skill) query.getSingleResult();
-                System.out.println("Habilidad " + results.getSkillName() + " con id " + results.getSkillId() + " eliminada con exito");
-                entityManager.remove(results);
+    transaction.begin();
 
-            }catch (NoResultException e){
-                System.out.println("No existe esa skill");
-                return;
-            }
+    Query query = entityManager.createNativeQuery("SELECT * FROM SKILL WHERE SKILL_NAME = ?", Skill.class)
+            .setParameter(1, skill);
 
-            //borrarla de la tabla skills
-            //borrarla de cualquier skillset que la contenga (iterar por los skillsets y si la contienen eliminarla)
-            transaction.commit();
+    Skill results = (Skill) query.getSingleResult();
+    entityManager.remove(results);
+    System.out.println("Habilidad " + results.getSkillName() + " con id " + results.getSkillId() + " eliminada con exito");
 
+
+    //borrarla de cualquier skillset que la contenga (iterar por los skillsets y si la contienen eliminarla)
+
+
+    query =entityManager.createNativeQuery("SELECT * FROM SKILLSET WHERE SKILL = ?").setParameter(1,results.getSkillId());
+    List<Skillset> remSS = query.getResultList();
+    for(Skillset i : remSS){
+        System.out.println(i.getSkill());
+        entityManager.remove(i);
     }
 
-    public static void addSkill(){
-
+        transaction.commit();
+        }else{
+    System.out.println("Esa skill no existe tontito");
+}
     }
 
+    public static void addSkill(int SS, int SK, int LVL){  //SS se obtiene del Employee, SK de Skill, y LVL es manual
+        transaction.begin();
+            Skillset skillset = new Skillset(highestID()[0],SS,SK,LVL );
 
-
-
+        transaction.commit();
+    }
     public static Integer iterator(ArrayList<Integer> arr){
         Integer max = arr.get(0);
         for(Integer i: arr){
